Add files via upload
Not sure I implemented this correctly. Everything I saw about 3.0 said the main difference is a timer on the send side.  So I copied the existing code into 3.0 and created a timer.  You will see my comments.  It worked for me (after a few revisions).  The client and server were updated to look for rdt_3_0.  Anyway, look it over and make sure it is correct.